requirements.txt
mysql-connector-python>=8.0.0
requests>=2.25.0
urllib3>=1.26.0
beautifulsoup4>=4.9.0
lxml>=4.6.0
opencv-python>=4.5.0
m3u8>=3.0.0
numpy>=1.20.0
timeout-decorator>=0.5.0
python-dotenv>=0.19.0
